Create DismissableTutorialMessage hook variation
